requirements.txt
faster-whisper>=1.0.0
openai>=1.0.0
pydub>=0.25.1
python-dotenv>=1.0.0